# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Aditya General Store API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: categoryId
          in: query
          required: false
          schema:
            type: integer
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /products/{id}/reviews:
    get:
      operationId: listProductReviews
      tags: [reviews]
      summary: List reviews for a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createProductReview
      tags: [reviews]
      summary: Create a review for a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /users:
    post:
      operationId: createUser
      tags: [users]
      summary: Create/login a user by phone number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "200":
          description: User (created or existing)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    Product:
      type: object
      required:
        - id
        - categoryId
        - name
        - description
        - price
        - mrp
        - unit
        - imageUrl
        - stock
        - lowStockThreshold
        - isBestSeller
        - isNewArrival
        - rating
        - reviewCount
        - createdAt
      properties:
        id: { type: integer }
        categoryId: { type: integer }
        name: { type: string }
        description: { type: string }
        price: { type: number }
        mrp: { type: number }
        unit: { type: string }
        imageUrl: { type: string }
        stock: { type: integer }
        lowStockThreshold: { type: integer }
        isBestSeller: { type: boolean }
        isNewArrival: { type: boolean }
        rating: { type: number }
        reviewCount: { type: integer }
        createdAt: { type: string, format: date-time }
    ProductInput:
      type: object
      required: [categoryId, name, description, price, mrp, unit, imageUrl, stock]
      properties:
        categoryId: { type: integer }
        name: { type: string, minLength: 1 }
        description: { type: string }
        price: { type: number }
        mrp: { type: number }
        unit: { type: string }
        imageUrl: { type: string }
        stock: { type: integer }
        lowStockThreshold: { type: integer }
        isBestSeller: { type: boolean }
        isNewArrival: { type: boolean }
    ProductUpdate:
      type: object
      properties:
        categoryId: { type: integer }
        name: { type: string, minLength: 1 }
        description: { type: string }
        price: { type: number }
        mrp: { type: number }
        unit: { type: string }
        imageUrl: { type: string }
        stock: { type: integer }
        lowStockThreshold: { type: integer }
        isBestSeller: { type: boolean }
        isNewArrival: { type: boolean }
    Review:
      type: object
      required: [id, productId, userName, rating, comment, createdAt]
      properties:
        id: { type: integer }
        productId: { type: integer }
        userName: { type: string }
        rating: { type: integer }
        comment: { type: string }
        createdAt: { type: string, format: date-time }
    ReviewInput:
      type: object
      required: [userName, rating, comment]
      properties:
        userName: { type: string, minLength: 1 }
        rating: { type: integer, minimum: 1, maximum: 5 }
        comment: { type: string }
    User:
      type: object
      required:
        - id
        - name
        - phone
        - loyaltyPoints
        - walletBalance
        - isFirstOrder
        - referralCode
        - createdAt
      properties:
        id: { type: integer }
        name: { type: string }
        phone: { type: string }
        loyaltyPoints: { type: integer }
        walletBalance: { type: number }
        isFirstOrder: { type: boolean }
        referralCode: { type: string }
        createdAt: { type: string, format: date-time }
    UserInput:
      type: object
      required: [name, phone]
      properties:
        name: { type: string, minLength: 1 }
        phone: { type: string, minLength: 1 }
    UserUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
...