# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Aditya General Store API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  categoryId: number;
  /** @minLength 1 */
  name: string;
  description: string;
  price: number;
  mrp: number;
  unit: string;
  imageUrl: string;
  stock: number;
  lowStockThreshold?: number;
  isBestSeller?: boolean;
  isNewArrival?: boolean;
}