# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Aditya General Store API specification
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  name: string;
  phone: string;
  loyaltyPoints: number;
  walletBalance: number;
  isFirstOrder: boolean;
  referralCode: string;
  createdAt: Date;
}